split models file in more files
