+++ hansberger/research/models/__init__.py
@@ -1,0 +1,2 @@
+from .research import Research
+from .dataset import Dataset
